backend/requirements.txt
langchain-text-splitters
langchain-google-genai
langchain-pinecone
pinecone
pymupdf
tqdm
dotenv
fastapi
uvicorn
pytesseract
pillow